Add union type-level support
