+++ src/utility/union/Union.ts
@@ -1,0 +1,3 @@
+export function Union<T extends any[]>(...args: T) {
+  return new UnionType(...args);
+}
