+++ src/utility/union/type/UnionType.ts
@@ -1,0 +1,7 @@
+class UnionType<T extends any[]> {
+  subtypes: T;
+
+  constructor(...args: T) {
+    this.subtypes = args;
+  }
+}
